Linux fixes for the first Host compile since HS-5/7: rd-spike's frame trampoline learns the graph_us the capture callback grew, and the beacon-offset literal picks Int64 explicitly for 6.1.2's inference

+++ Host/Tests/HostWireTests/SessionGateTests.swift
@@ -496,7 +496,7 @@ final class SessionGateTests: XCTestCase {
         )
         XCTAssertEqual(echoEvents, [.beaconEchoAccepted(
             beaconSeq: 0,
-            offsetMicroseconds: (3_000 + (-3_500)) / 2,
+            offsetMicroseconds: Int64(3_000 + (-3_500)) / 2,
             rttMicroseconds: 6_500
         )])
         XCTAssertEqual(session.clock.samples, 1)
